requirements.txt
google-cloud-logging
google-cloud-storage
canvas-data-sdk